+++ benchmarks/Classifiers/RandomForestBench.php
@@ -13,7 +13,6 @@ use Rubix\ML\Transformers\IntervalDiscretizer;
 use Rubix\ML\Backends\Serial;
 use Rubix\ML\Backends\Amp;
 use Rubix\ML\Backends\Swoole;
-use Rubix\ML\Specifications\SpecificationChain;
 use Rubix\ML\Specifications\ExtensionIsLoaded;
 
 /**
@@ -48,12 +47,7 @@ class RandomForestBench
             'backend' => $ampBackend,
         ];
 
-        if (
-            SpecificationChain::with([
-                new ExtensionIsLoaded('swoole'),
-                new ExtensionIsLoaded('igbinary'),
-            ])->passes()
-        ) {
+        if (ExtensionIsLoaded::with('swoole')->passes()) {
             $swooleBackend = new Swoole();
 
             yield (string) $swooleBackend => [